requirements.txt
django >=3.2,<4.0
gunicorn
psycopg2-binary
django-dotenv
pint
django-htmx
django-mptt
django-forms-dynamic
django-widget-tweaks
django-crispy-forms
crispy-bootstrap5
django-formtools